all vector variants and refactoring of folder structure
