+++ operators/src/processing/expression/raster_query_processor.rs
