+++ operators/src/processing/mod.rs
@@ -17,13 +17,15 @@ mod rgb;
 mod temporal_raster_aggregation;
 mod time_projection;
 mod time_shift;
-mod vector_expression;
 mod vector_join;
 
 pub use circle_merging_quadtree::{
     InitializedVisualPointClustering, VisualPointClustering, VisualPointClusteringParams,
 };
-pub use expression::{Expression, ExpressionError, ExpressionParams};
+pub use expression::{
+    Expression, ExpressionError, ExpressionParams, VectorExpression, VectorExpressionError,
+    VectorExpressionParams,
+};
 pub use interpolation::{Interpolation, InterpolationError, InterpolationParams};
 pub use line_simplification::{
     LineSimplification, LineSimplificationError, LineSimplificationParams,
@@ -52,4 +54,3 @@ pub use temporal_raster_aggregation::{
 };
 pub use time_projection::{TimeProjection, TimeProjectionError, TimeProjectionParams};
 pub use time_shift::{TimeShift, TimeShiftError, TimeShiftParams};
-pub use vector_expression::{VectorExpression, VectorExpressionError, VectorExpressionParams};
